add *_vc17 project and solution by VC17 migration from _vc9 solution

+++ cmd_line/TransferFiles/FileTransfer.cpp
@@ -74,7 +74,7 @@ int CFileTransfer::Transfer( void )
 		}
 	}
 
-	return m_fileCount;
+	return static_cast<int>( m_fileCount );
 }
 
 void CFileTransfer::SearchSourceFiles( const fs::CPath& dirPath )
